+++ Lib/Conf/confMY.php
@@ -3,7 +3,7 @@
     $server="127.0.0.1";//es como la dirreccion y ip del localost, como por defecto
     $user="root";
     $pass=""; //contraseña de la base de datos de postgres
-    $database="vialmanager";//Colocar el nombre respectivo de la BD
+    $database="vial_manager2_6";//Colocar el nombre respectivo de la BD
     //PD: juan sebastian no recuerda la version de la base de datos de mysql XD
 
     $port="3306";//5432
